feat: moved unit tests files under vcdm
