+++ packages/core/tests/vcdm/abi/fixture.ts
@@ -1,7 +1,7 @@
 import { InvalidAbiDataToEncodeOrDecode } from '@vechain/sdk-errors';
-import { Address, HexUInt } from '../../src';
-import { Hex } from '../../src/vcdm/Hex';
-import { generateRandomValidAddress } from '../fixture';
+import { Address, HexUInt } from '../../../src';
+import { Hex } from '../../../src/vcdm/Hex';
+import { generateRandomValidAddress } from '../../fixture';
 
 /**
  * Simple functions fixtures
